# backend/tests/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryCreate": {
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "slug": {
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "CategoryCreate",
        "type": "object"
      },
      "CategoryResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "slug": {
            "maxLength": 120,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug",
          "id"
        ],
        "title": "CategoryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductCreate": {
        "properties": {
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "default_unit": {
            "enum": [
              "kg",
              "g",
              "crate",
              "piece",
              "bunch",
              "bag"
            ],
            "title": "Default Unit",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "maxLength": 160,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "category_id",
          "name",
          "slug",
          "default_unit"
        ],
        "title": "ProductCreate",
        "type": "object"
      },
      "ProductResponse": {
        "properties": {
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "default_unit": {
            "enum": [
              "kg",
              "g",
              "crate",
              "piece",
              "bunch",
              "bag"
            ],
            "title": "Default Unit",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "maxLength": 160,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "category_id",
          "name",
          "slug",
          "default_unit",
          "id"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Mavuno API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/catalog/categories": {
      "get": {
        "operationId": "categories_api_v1_catalog_categories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  },
                  "title": "Response Categories Api V1 Catalog Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Categories",
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "create_category_api_v1_catalog_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Category",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/products": {
      "get": {
        "operationId": "products_api_v1_catalog_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  },
                  "title": "Response Products Api V1 Catalog Products Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Products",
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "create_product_api_v1_catalog_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Product",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/farmers/me/listings": {
      "get": {
        "operationId": "owned_listings_api_v1_farmers_me_listings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ListingResponse"
                  },
                  "title": "Response Owned Listings Api V1 Farmers Me Listings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Owned Listings",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/farmers/{farmer_id}": {
      "get": {
        "operationId": "farmer_profile_api_v1_farmers__farmer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "farmer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Farmer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerPublicProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Farmer Profile",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/listing-images/{image_id}": {
      "get": {
        "operationId": "listing_photo_api_v1_listing_images__image_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing Photo",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings": {
      "get": {
        "operationId": "list_listings_api_v1_listings_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "minLength": 2,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "farmer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Farmer Id"
            }
          },
          {
            "in": "query",
            "name": "unit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "kg",
                    "g",
                    "crate",
                    "piece",
                    "bunch",
                    "bag"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit"
            }
          },
          {
            "in": "query",
            "name": "min_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Price"
            }
          },
          {
            "in": "query",
            "name": "max_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "newest",
              "enum": [
                "newest",
                "price_asc",
                "price_desc"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Listings",
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "create_listing_api_v1_listings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Listing",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}": {
      "delete": {
        "operationId": "archive_listing_api_v1_listings__listing_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Expected Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Listing",
        "tags": [
          "catalog"
        ]
      },
      "get": {
        "operationId": "get_listing_api_v1_listings__listing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Listing",
        "tags": [
          "catalog"
        ]
      },
      "patch": {
        "operationId": "update_listing_api_v1_listings__listing_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Listing",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/images": {
      "post": {
        "operationId": "add_image_api_v1_listings__listing_id__images_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Image",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/images/{image_id}": {
      "delete": {
        "operationId": "delete_image_api_v1_listings__listing_id__images__image_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Image",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/inventory": {
      "post": {
        "operationId": "change_inventory_api_v1_listings__listing_id__inventory_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Inventory",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/photos": {
      "post": {
        "operationId": "upload_photo_api_v1_listings__listing_id__photos_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Photo",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}